[increment-5] fix reviewer remarks #3
